breaking compatibility with the https://github.com/Heavy-Vehicle-Networking-At-U-Tulsa/Log-File-Format-Converter to provide an easy to use wrapper in the spirit of j4l's changes, but maintainin the construction of the describer in different levels of abstraction
