+++ pretty_j1939.py
@@ -5,60 +5,67 @@ import argparse
 import sys
 import json
 
-from pretty_j1939.prettify import Prettyfier
+import pretty_j1939.describe
 
 
 parser = argparse.ArgumentParser(description='pretty-printing J1939 candump logs')
 parser.add_argument('candump', help='candump log')
 
-parser.add_argument('--da-json', type=str, const=True, default="J1939db.json", nargs='?',
+parser.add_argument('--da-json', type=str, const=True, default=pretty_j1939.describe.DEFAULT_DA_JSON, nargs='?',
                     help='absolute path to the input JSON DA (default=\"./J1939db.json\")')
 
 parser.add_argument('--candata',    dest='candata', action='store_true',  help='print input can data')
 parser.add_argument('--no-candata', dest='candata', action='store_false', help='(default)')
-parser.set_defaults(candata=False)
+parser.set_defaults(candata=pretty_j1939.describe.DEFAULT_CANDATA)
 
 parser.add_argument('--pgn',    dest='pgn', action='store_true', help='(default) print source/destination/type '
                                                                       'description')
 parser.add_argument('--no-pgn', dest='pgn', action='store_false')
-parser.set_defaults(pgn=True)
+parser.set_defaults(pgn=pretty_j1939.describe.DEFAULT_PGN)
 
 parser.add_argument('--spn',    dest='spn', action='store_true', help='(default) print signals description')
 parser.add_argument('--no-spn', dest='spn', action='store_false')
-parser.set_defaults(spn=True)
+parser.set_defaults(spn=pretty_j1939.describe.DEFAULT_SPN)
 
 parser.add_argument('--transport',    dest='transport', action='store_true',  help='print details of transport-layer '
                                                                                    'streams found')
-parser.add_argument('--no-transport', dest='transport', action='store_false', help='(default)')
-parser.set_defaults(transport=False)
+parser.add_argument('--no-transport', dest='transport', action='store_false', help='(default)') #TODO valid single-frame J1939 is transport also
+parser.set_defaults(transport=pretty_j1939.describe.DEFAULT_TRANSPORT)
 
-parser.add_argument('--link',    dest='link', action='store_true', help='(default) print details of link-layer frames '
-                                                                        'found')
+parser.add_argument('--link',    dest='link', action='store_true', help='(default) print details of link-layer frames ' 
+                                                                        'found') # TODO as-above, valid single-frame J1939 is not link-layer
 parser.add_argument('--no-link', dest='link', action='store_false')
-parser.set_defaults(link=True)
+parser.set_defaults(link=pretty_j1939.describe.DEFAULT_LINK)
 
 parser.add_argument('--include_na',    dest='include_na', action='store_true',  help='include not-available (0xff) SPN '
                                                                                      'values')
 parser.add_argument('--no-include_na', dest='include_na', action='store_false', help='(default)')
 parser.set_defaults(include_na=False)
 
+parser.add_argument('--real-time',    dest='real_time', action='store_true',  help='emit SPNs as they are seen in '
+                                                                                   'transport sessions')
+parser.add_argument('--no-real-time', dest='real_time', action='store_false', help='(default)')
+parser.set_defaults(real_time=pretty_j1939.describe.DEFAULT_REAL_TIME)
+
 parser.add_argument('--format',    dest='format', action='store_true',  help='format each structure (otherwise '
                                                                              'single-line)')
 parser.add_argument('--no-format', dest='format', action='store_false', help='(default)')
 parser.set_defaults(format=False)
 
-parser.add_argument('--real-time',    dest='real_time', action='store_true',  help='emit SPNs as they are seen in '
-                                                                                   'transport sessions')
-parser.add_argument('--no-real-time', dest='real_time', action='store_false', help='(default)')
-parser.set_defaults(real_time=False)
 
 args = parser.parse_args()
 
+
 if __name__ == '__main__':
-    prettyfier = Prettyfier(args.da_json, real_time=args.real_time, describe_pgns=args.pgn, describe_spns=args.spn,
-                            describe_link_layer=args.link, describe_transport_layer=args.transport,
-                            include_transport_rawdata=args.candata,
-                            include_na=args.include_na)
+    describe = pretty_j1939.describe.get_describer(
+                                da_json=args.da_json,
+                                describe_pgns=args.pgn,
+                                describe_spns=args.spn,
+                                describe_link_layer=args.link,
+                                describe_transport_layer=args.transport,
+                                real_time=args.real_time,
+                                include_transport_rawdata=args.candata,
+                                include_na=args.include_na)
     with open(args.candump, 'r') as f:
         for candump_line in f.readlines():
             if candump_line == '\n':
@@ -75,7 +82,7 @@ if __name__ == '__main__':
 
             desc_line = ''
 
-            description = prettyfier.describer(message_data.bytes, message_id.uint)
+            description = describe(message_data.bytes, message_id.uint)
             if args.format:
                 json_description = str(json.dumps(description, indent=4))
             else:
